move display_outlines from shapes/draw_newshape to shapes
